Create SQL_Basics
On the basis of merit, a company decides to promote some of its employee in its HR division at the end of the quarter because of their high performance. Write a query to fin the employee IDs along with the name of all its employees who work in the HR department who earned a bonus of 5000 dollars or more in the last quarter.
There are two tables in the database employee_information and last_quarter_bonus. Their primary keys are employee_id.